wm8iGHFA - Split paginatedResult and findPaginated into two separate files
