+++ lib/http/paginatedResult.js
@@ -1,0 +1,9 @@
+class PaginatedResult {
+  constructor (items, currentPage, totalPages) {
+    this.items = items
+    this.currentPage = currentPage
+    this.totalPages = totalPages
+  }
+}
+
+module.exports = PaginatedResult
