+++ lib/http/paginator.js
@@ -1,3 +1,5 @@
+const PaginatedResult = require('./paginatedResult')
+
 /**
  * @param requestQuery {@link express~Request#query}
  * @see mongorito~Model.find
@@ -25,15 +27,4 @@ async function findPaginated (model, requestQuery, query) {
   return new PaginatedResult(result, Math.ceil(offset / count), Math.ceil(totalItems / count))
 }
 
-class PaginatedResult {
-  constructor (items, currentPage, totalPages) {
-    this.items = items
-    this.currentPage = currentPage
-    this.totalPages = totalPages
-  }
-}
-
-module.exports = {
-  findPaginated,
-  PaginatedResult
-}
+module.exports = findPaginated
